Revert "Merge branch 'master' into sdt_dd"
This reverts commit e4890adcddba54719be6f4efcb7e4e254c9f214f.

+++ media/webrtc/trunk/webrtc/modules/audio_coding/neteq/test/NETEQTEST_RTPpacket.h
@@ -36,12 +36,13 @@ public:
     int readFixedFromFile(FILE *fp, size_t len);
     virtual int writeToFile(FILE *fp);
     void blockPT(uint8_t pt);
+    //int16_t payloadType();
     virtual void parseHeader();
     void parseHeader(webrtc::WebRtcRTPHeader* rtp_header);
     const webrtc::WebRtcRTPHeader* RTPinfo() const;
     uint8_t * datagram() const;
     uint8_t * payload() const;
-    size_t payloadLen();
+    int16_t payloadLen();
     int16_t dataLen() const;
     bool isParsed() const;
     bool isLost() const;
@@ -72,7 +73,7 @@ public:
     uint8_t *       _payloadPtr;
     int                 _memSize;
     int16_t         _datagramLen;
-    size_t          _payloadLen;
+    int16_t         _payloadLen;
     webrtc::WebRtcRTPHeader _rtpInfo;
     bool                _rtpParsed;
     uint32_t        _receiveTime;
